Undo refactor of ElementDetail now that ElementCard will be used
